+++ GardenOfDreamsTestTask/Assets/Materials/Scripts/Player/IPlayerData.cs
@@ -11,10 +11,10 @@ public interface IPlayerData
     void SetWeight(float value);
 
     float GetHeadArmor();
-    void SetHeadArmor(int value);
+    void SetHeadArmor(float value);
 
     float GetBodyArmor();
-    void SetBodyArmor(int value);
+    void SetBodyArmor(float value);
 
     int GetMoney();
     void SetMoney(int value);
